move web3Proxy to unlock.js/
